+++ common/enhancedportals/lib/GuiIds.java
@@ -6,4 +6,5 @@ public class GuiIds
     public static final int PortalModifierNetwork = 2;
     public static final int DialDevice = 3;
     public static final int DialDeviceBasic = 4;
+    public static final int AutoDiallerBasic = 5;
 }