+++ common/enhancedportals/network/CommonProxy.java
@@ -6,6 +6,7 @@ import net.minecraft.world.World;
 import net.minecraftforge.common.Configuration;
 import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.registry.GameRegistry;
+import enhancedportals.block.BlockAutomaticDialler;
 import enhancedportals.block.BlockDialDevice;
 import enhancedportals.block.BlockDialDeviceBasic;
 import enhancedportals.block.BlockDummyPortal;
@@ -21,6 +22,7 @@ import enhancedportals.lib.Localization;
 import enhancedportals.lib.Settings;
 import enhancedportals.portal.network.DialDeviceNetwork;
 import enhancedportals.portal.network.ModifierNetwork;
+import enhancedportals.tileentity.TileEntityAutomaticDialler;
 import enhancedportals.tileentity.TileEntityDialDevice;
 import enhancedportals.tileentity.TileEntityDialDeviceBasic;
 import enhancedportals.tileentity.TileEntityNetherPortal;
@@ -35,6 +37,7 @@ public class CommonProxy
     public BlockDummyPortal blockDummyPortal;
     public BlockDialDevice blockDialDevice;
     public BlockDialDeviceBasic blockDialDeviceBasic;
+    public BlockAutomaticDialler blockAutomaticDialler;
 
     public ItemPortalModifierUpgrade portalModifierUpgrade;
     public ItemNetworkCard networkCard;
@@ -58,6 +61,7 @@ public class CommonProxy
         blockDummyPortal = new BlockDummyPortal();
         blockDialDevice = new BlockDialDevice();
         blockDialDeviceBasic = new BlockDialDeviceBasic();
+        blockAutomaticDialler = new BlockAutomaticDialler();
 
         GameRegistry.registerBlock(blockNetherPortal, Localization.NetherPortal_Name);
         GameRegistry.registerBlock(blockPortalModifier, Localization.PortalModifier_Name);
@@ -66,6 +70,7 @@ public class CommonProxy
         GameRegistry.registerBlock(blockDummyPortal, "Fake Portal");
         GameRegistry.registerBlock(blockDialDevice, Localization.DialDevice_Name);
         GameRegistry.registerBlock(blockDialDeviceBasic, Localization.DialDeviceBasic_Name);
+        GameRegistry.registerBlock(blockAutomaticDialler, Localization.AutomaticDialler_Name);
     }
 
     public void loadItems()
@@ -118,5 +123,6 @@ public class CommonProxy
         GameRegistry.registerTileEntity(TileEntityPortalModifier.class, "EPPModifier");
         GameRegistry.registerTileEntity(TileEntityDialDevice.class, "EPDDevice");
         GameRegistry.registerTileEntity(TileEntityDialDeviceBasic.class, "EPDDBasic");
+        GameRegistry.registerTileEntity(TileEntityAutomaticDialler.class, "EPADialler");
     }
 }